Refactor API Make api testable Write first test for API
